REFACTOR: refactor test for non ocr pdf extraction
